Moved some tests into their own files
